+++ src/components/events/index.jsx
@@ -1,0 +1,3 @@
+export * from './events.jsx';
+export * from './slides.jsx';
+export * from './event.jsx';
